fix: 타입 에러 수정 및 칼럼 오류 수정

+++ backend/src/main/java/com/example/Piroin/project/domain/deposit/service/DepositService.java
@@ -20,7 +20,7 @@ public class DepositService {
     private final AttendanceRepository attendanceRepository;
 
     @Transactional
-    public void recalculateDeposit(Long userId) {
+    public void recalculateDeposit(Integer userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 유저입니다."));
